Fix getchar() prototype definition in test program
According to the SDCC manual (sdccman.pdf), the getchar() prototype
changed in version 3.7.0. Based on the code/comment that is being
replaced in this commit, only versions between 3.6.5 and 3.7.0 use the
char return type. Older versions than 3.6.5 use unsigned char (and that
is still supported), and newer versions than 3.7.0 use int.

This is related to commit 28639fc7dd6a170436201f0e5b7ea3a4928a3cf9, but
fixes SDCC compatibility correctly. The whole point is to redefine the
getchar() function and make the redefined version available to the whole
C library (as opposed to defining a private function that works only in
the test program when it's called directly).