Fix board detection in test program
The test program Makefile defines DEVICE=STM8xxxx but only internally.
The argument that is passed to the C compiler is -D$(DEVICE), which
means at the C preprocessor level only the STM8xxxx part is available.
In other words, we have a STM8xxxx macro with an empty value instead of
a DEVICE macro with a STM8xxxx value.